CLSO-2670: Deduplicate leapp upgrader code and drop centos support
cloudlinux elevation code will be in another repository, so
there is no need to keep centos support
