+++ centos2almaconverter/actions/installation.py
@@ -4,7 +4,6 @@ import shutil
 
 from pleskdistup.common import action, rpm, util
 
-LEAPP_ALMALINUX_RPM_URL = "https://repo.almalinux.org/elevate/elevate-release-latest-el7.noarch.rpm"
 LEAPP_CLOUDLINUX_RPM_URL = "https://repo.cloudlinux.com/elevate/elevate-release-latest-el7.noarch.rpm"
 
 
@@ -13,11 +12,11 @@ class LeapInstallation(action.ActiveAction):
     def __init__(self, elevate_release_rpm_url, pkgs_to_install):
         self.name = "installing leapp"
         self.pkgs_to_install = pkgs_to_install
-        self._elevate_release_rpm_url = elevate_release_rpm_url
+        self.elevate_release_rpm_url = elevate_release_rpm_url
 
     def _prepare_action(self) -> action.ActionResult:
         if not rpm.is_package_installed("elevate-release"):
-            util.logged_check_call(["/usr/bin/yum", "install", "-y", "https://repo.almalinux.org/elevate/elevate-release-latest-el7.noarch.rpm"])
+            util.logged_check_call(["/usr/bin/yum", "install", "-y", self.elevate_release_rpm_url])
 
         util.logged_check_call(["/usr/bin/yum-config-manager", "--enable", "elevate"])
