+++ centos2almaconverter/main.py
@@ -7,9 +7,7 @@ import pleskdistup.main
 import pleskdistup.registry
 
 import centos2almaconverter.upgrader
-import centos2almaconverter.cl_upgrader
 
 if __name__ == "__main__":
     pleskdistup.registry.register_upgrader(centos2almaconverter.upgrader.Centos2AlmaConverterFactory())
-    pleskdistup.registry.register_upgrader(centos2almaconverter.cl_upgrader.CloudLinuxConverterFactory())
     sys.exit(pleskdistup.main.main())
